# cmd/web/docs/swagger.yaml
basePath: /
definitions:
  categories.CreateCategoryRequest:
    properties:
      name:
        type: string
      user_id:
        type: integer
    type: object
  categories.CreateCategoryUpdateRequest:
    properties:
      name:
        type: string
      user_id:
        type: integer
    type: object
  common.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  common.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  users.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  users.CreateUserUpdateRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
  title: sprig API
  version: "1.0"
paths:
  /api/v1/categories:
    get:
      description: Mengembalikan daftar seluruh kategori
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Ambil semua kategori
      tags:
      - categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Data kategori baru
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/categories.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Tambah category baru
      tags:
      - categories
  /api/v1/categories/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID Kategori
        in: path
        name: id
        required: true
        type: integer
      - description: Field yang ingin diupdate
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/categories.CreateCategoryUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update sebagian data Kategori
      tags:
      - categories
  /api/v1/categories/{name}:
    delete:
      parameters:
      - description: Nama Kategori
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Hapus kategori
      tags:
      - categories
    get:
      parameters:
      - description: Nama Kategori
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Ambil detail kategori berdasarkan Nama
      tags:
      - categories
  /api/v1/users:
    get:
      description: Mengembalikan daftar seluruh user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Ambil semua user
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: Data user baru
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Tambah user baru
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Hapus user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Ambil detail user berdasarkan ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Field yang ingin diupdate
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update sebagian data user
      tags:
      - users
swagger: "2.0"